requirements.txt
absl-py==0.6.1
admin-totals==1.0.1
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
arrow==0.13.0
astor==0.7.1
bcrypt==3.1.6
billiard==3.3.0.23
binaryornot==0.4.4
cached-property==1.5.1
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.7
cookiecutter==1.6.0
coverage==4.3.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.13
django-admin-rangefilter==0.3.12
django-advanced-filters==1.1.1
django-allauth==0.38.0
django-background-tasks==1.2.0
django-braces==1.13.0
django-celery==3.2.2
django-compat==1.0.15
django-crispy-forms==1.7.2
django-crum==0.7.3
django-currentuser==0.2.3
django-extensions==2.1.4
django-filter==2.0.0
django-hashers-passlib==0.4
django-merged-inlines==1.0.3
django-permissions==1.2.2
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-url-filter==0.3.11
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-nested-routers==0.91
drf-writable-nested==0.5.1
enum-compat==0.0.2
flask-spawn==0.1.11
future==0.17.1
gast==0.2.0
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.8.0
httpie==1.0.3
httplib2==0.12.0
idna==2.8
Jinja2==2.10
jinja2-time==0.2.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kombu==3.0.37
Markdown==3.0.1
MarkupSafe==0.23
numpy==1.15.4
oauth2client==4.1.3
oauthlib==2.1.0
packaging==16.8
pandas==0.23.4
passlib==1.7.1
pbr==5.1.1
Pillow==6.2.0
poyo==0.4.2
protobuf==3.6.1
psycopg2==2.7.6.1
py==1.4.32
pyad==0.5.20
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pyflakes==1.5.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.1.10
pytest==3.0.6
python-dateutil==2.6.0
python-decouple==3.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
RestAuth==0.6.3
RestAuthCommon==0.7.0
rsa==4.0
simplejson==3.10.0
six==1.10.0
South==1.0.2
stevedore==1.30.0
termcolor==1.1.0
urllib3==1.24.2
vine==1.1.4
virtualenv==15.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
Werkzeug==0.15.3
whichcraft==0.5.2
whitenoise==3.2.1
